+++ SampleApp/NotebookTableViewController.h
@@ -1,0 +1,15 @@
+//
+//  NotebookTableViewController.h
+//  evernote-sdk-ios
+//
+//  Created by Mustafa Furniturewala on 6/14/13.
+//  Copyright (c) 2013 n/a. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface NotebookTableViewController : UITableViewController
+
+@property (nonatomic,strong) NSArray* notebooks;
+
+@end